Revert "Revert "Merge pull request #91 from aura-systems/master""
This reverts commit e364b88ebf68c073e03b513f8d237ce2467f21d3.

+++ Aura Operating System/Aura OS/System/Utils/Settings.cs
@@ -26,7 +26,10 @@ namespace Aura_OS.System.Utils
             //reset of config in memory if there is "something"
             file = reset;
             //load
-            file = File.ReadAllLines(@"0:\System\settings.conf");
+            if (Kernel.SystemExists)
+            {
+                file = File.ReadAllLines(@"0:\System\settings.conf");
+            }
         }
 
         /// <summary>
@@ -34,7 +37,10 @@ namespace Aura_OS.System.Utils
         /// </summary>
         public static void PushValues()
         {
-            File.WriteAllLines(@"0:\System\settings.conf", file);
+            if (Kernel.SystemExists)
+            {
+                File.WriteAllLines(@"0:\System\settings.conf", file);
+            }
         }
 
         /// <summary>
@@ -68,6 +74,12 @@ namespace Aura_OS.System.Utils
         /// </summary>
         public static string GetValue(string parameter)
         {
+
+            if (file == null)
+            {
+                return "null";
+            }
+
             string value = "null";
 
             foreach (string line in file)
